requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
SQLAlchemy>=2.0.0
aiosqlite>=0.17.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai-whisper>=20231117
loguru>=0.7.0
alembic>=1.12.0
asyncio>=3.4.3
aiohttp>=3.8.1 